requirements.txt
requests==2.32.3
tenacity==9.0.0
jsonschema==4.23.0
Pillow==11.0.0
python-dotenv==1.0.1
